feat(connection): Enforce deprecation of old con. api by module design
Move deprecated connection modules to legacy package and remove rework package to allow for one final breaking change once the new connection api is finished.
