+++ modules/telestion-services/src/main/java/de/wuespace/telestion/services/connection/RawMessage.java
@@ -3,11 +3,8 @@ package de.wuespace.telestion.services.connection;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.wuespace.telestion.api.message.JsonMessage;
 
-@Deprecated(since = "v0.1.3", forRemoval = true)
-public record SerialData(@JsonProperty byte[] data) implements JsonMessage {
-
-	@SuppressWarnings("unused")
-	private SerialData(){
+public record RawMessage(@JsonProperty byte[] data) implements JsonMessage {
+	private RawMessage() {
 		this(null);
 	}
 }
