--- modules/telestion-services/src/main/java/de/wuespace/telestion/services/connection/rework/RawMessage.java
@@ -1,10 +1,0 @@
-package de.wuespace.telestion.services.connection.rework;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import de.wuespace.telestion.api.message.JsonMessage;
-
-public record RawMessage(@JsonProperty byte[] data) implements JsonMessage {
-	private RawMessage() {
-		this(null);
-	}
-}
