+++ modules/telestion-services/src/main/java/de/wuespace/telestion/services/connection/tcp/TcpDetails.java
@@ -1,7 +1,7 @@
-package de.wuespace.telestion.services.connection.rework.tcp;
+package de.wuespace.telestion.services.connection.tcp;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import de.wuespace.telestion.services.connection.rework.ConnectionDetails;
+import de.wuespace.telestion.services.connection.ConnectionDetails;
 
 public record TcpDetails(@JsonProperty String ip,
 						 @JsonProperty int port,
