+++ modules/telestion-services/src/main/java/de/wuespace/telestion/services/connection/tcp/TcpDispatcher.java
@@ -1,11 +1,11 @@
-package de.wuespace.telestion.services.connection.rework.tcp;
+package de.wuespace.telestion.services.connection.tcp;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import de.wuespace.telestion.api.config.Config;
 import de.wuespace.telestion.api.message.JsonMessage;
-import de.wuespace.telestion.services.connection.rework.ConnectionData;
-import de.wuespace.telestion.services.connection.rework.SenderData;
-import de.wuespace.telestion.services.connection.rework.Tuple;
+import de.wuespace.telestion.services.connection.ConnectionData;
+import de.wuespace.telestion.services.connection.SenderData;
+import de.wuespace.telestion.services.connection.Tuple;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.Promise;
