+++ modules/telestion-services/src/main/java/de/wuespace/telestion/services/connection/udp/UdpDetails.java
@@ -1,7 +1,7 @@
-package de.wuespace.telestion.services.connection.rework.udp;
+package de.wuespace.telestion.services.connection.udp;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import de.wuespace.telestion.services.connection.rework.ConnectionDetails;
+import de.wuespace.telestion.services.connection.ConnectionDetails;
 
 public record UdpDetails(@JsonProperty String ip,
 						 @JsonProperty int port,
